// src/generated/prisma/enums.ts
export const UserRole = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  ADMIN: 'ADMIN',
  OWNER: 'OWNER',
  EMPLOYEE: 'EMPLOYEE',
  SUPPLIER: 'SUPPLIER',
  CUSTOMER: 'CUSTOMER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]